requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
dnspython==1.16.0
forex-python==1.5
idna==2.9
importlib-metadata==1.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.8.2
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-engineio==3.13.0
python-socketio==4.6.0
requests==2.24.0
simplejson==3.17.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.4
zipp==3.1.0
